+++ Source/Core/Core/PowerPC/Jit64IL/JitIL.h
@@ -30,7 +30,6 @@
 #include "Core/PowerPC/PPCTables.h"
 #include "Core/PowerPC/Jit64/JitAsm.h"
 #include "Core/PowerPC/JitCommon/Jit_Util.h"
-#include "Core/PowerPC/JitCommon/JitBackpatch.h"
 #include "Core/PowerPC/JitCommon/JitBase.h"
 #include "Core/PowerPC/JitCommon/JitCache.h"
 #include "Core/PowerPC/JitILCommon/IR.h"
