+++ Source/Core/Core/PowerPC/JitArm32/JitArm_BackPatch.cpp
@@ -9,7 +9,6 @@
 
 #include "Core/HW/Memmap.h"
 #include "Core/PowerPC/JitArm32/Jit.h"
-#include "Core/PowerPC/JitCommon/JitBackpatch.h"
 
 using namespace ArmGen;
